% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIPAC.R
\name{classifier.Lambda.core}
\alias{classifier.Lambda.core}
\title{The core function to obtain Lambda using different regression methods}
\usage{
classifier.Lambda.core(bulk.dat, y, family, K.means.res, ela.net.alpha = 0.4)
}
\arguments{
\item{bulk.dat}{a dimension-reduced bulk data. Rows stand for samples and columns stand for PCs.}

\item{y}{class labels or survival time.
If family = "binomial", "cumulative", or "gaussian", y represents class labels.
If family = "cox", y is a data frame with two columns. Row names of y are the same as bulk.dat's and the two columns are named ("time", "status").
This renaming is required for the algorithm used.}

\item{family}{this argument takes one of the following input:
(1). "binomial": logistic regression with elastic net;
(2). "cumulative": reversed proportional log odds ratio model with elastic net;
(3). "gaussian": linear regression with elastic net; and
(4). "cox": cox regression with elastic net.}

\item{K.means.res}{a list contains (1). k, the number of clusters;
(2). ct.assignment, a data frame with one column indicating cluster assignment. Row names of ct.assignment are cell names
(3). centers, cluster centroids. Rows are for PCs and columns are for clusters.
Each cluster centroid is calculated by taking the average value of all the cells in the cluster.}

\item{ela.net.alpha}{the parameter alpha used for the elastic net. 0 for ridge and 1 for lasso. The default is 0.4.}
}
\value{
a data frame with two columns. One column named ct.assignment is the cluster assignment, and the other one named
Lambda is the calculated Lambda for each cell. Cells from the same cluster have the same value of Lambda.
}
\description{
The core function to obtain Lambda using different regression methods
}
