% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIPAC.R
\name{obtain.ct.Lambda}
\alias{obtain.ct.Lambda}
\title{Summarize Lambda values}
\usage{
obtain.ct.Lambda(Lambda.res, K.means.res, CI.alpha = 0.05)
}
\arguments{
\item{Lambda.res}{Lambda values for each cell with multiple bootstrap samples.}

\item{K.means.res}{a list contains (1). k, the number of clusters;
(2). ct.assignment, a data frame with one column indicating cluster assignment. Row names of ct.assignment are cell names
(3). centers, cluster centroids. Rows are for PCs and columns are for clusters.
Each cluster centroid is calculated by taking the average value of all the cells in the cluster.}

\item{CI.alpha}{significance level used to decide significantly positive/negative results. The default is 0.05.}
}
\value{
A data frame whose rows are cells and columns are bootstrap samples.
}
\description{
Summarize Lambda values
}
