% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.sc.bulk.dat.R
\name{obtain.preprocessed.spatial.data}
\alias{obtain.preprocessed.spatial.data}
\title{Normalize spatial data and find highly variable genes using Seurat}
\usage{
obtain.preprocessed.spatial.data(spatial.obj, hvg = 1000, assay = NULL)
}
\arguments{
\item{spatial.obj}{Seurat spatial object (e.g., from Load10X_Spatial)}

\item{hvg}{the number of highly variable genes. The default is 1000.}

\item{assay}{the assay to use. If NULL, will auto-detect from available assays.
Common values: "Spatial" (Visium), "SCT" (spatial SCTransform), "RNA" (regular spatial)}
}
\value{
a spatial expression matrix. Rows stand for genes and columns stand for spots/cells.
}
\description{
Normalize spatial data and find highly variable genes using Seurat
}
