% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.sc.bulk.dat.R
\name{preprocess.sc.bulk.dat}
\alias{preprocess.sc.bulk.dat}
\title{Pre-process both single-cell and bulk data}
\usage{
preprocess.sc.bulk.dat(sc.dat, bulk.dat, hvg = 1000)
}
\arguments{
\item{sc.dat}{single-cell data matrix. Rows stand for genes and columns stand for cells.}

\item{bulk.dat}{bulk data matrix. Rows stand for genes and columns stand for cells.}

\item{hvg}{the number of highly variable genes. The default is \code{1000}.}
}
\value{
A list object contains pre-processed single-cell and bulk data
}
\description{
This function first finds overlap genes between scRNA-seq and bulk data,
then normalizes scRNA-seq data and find variable genes,
log-transforms bulk RNA-seq data,
and uses high variable genes for both scRNA-seq and bulk data
}
