% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.sc.bulk.dat.R
\name{preprocess.spatial.bulk.dat}
\alias{preprocess.spatial.bulk.dat}
\title{Pre-process both spatial transcriptomics and bulk data}
\usage{
preprocess.spatial.bulk.dat(spatial.obj, bulk.dat, hvg = 1000, assay = NULL)
}
\arguments{
\item{spatial.obj}{Seurat spatial object (e.g., from Load10X_Spatial)}

\item{bulk.dat}{bulk data matrix. Rows stand for genes and columns stand for samples.}

\item{hvg}{the number of highly variable genes. The default is \code{1000}.}

\item{assay}{the assay to use from spatial object. If NULL, will auto-detect.}
}
\value{
A list object contains pre-processed spatial and bulk data
}
\description{
This function first finds overlap genes between spatial and bulk data,
then normalizes spatial data and finds variable genes,
log-transforms bulk RNA-seq data,
and uses high variable genes for both spatial and bulk data
}
