% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIPAC.R
\name{SCIPAC}
\alias{SCIPAC}
\title{Obtain phenotype-associated cells using SCIPAC}
\usage{
SCIPAC(
  bulk.dat,
  y,
  family,
  ct.res,
  ela.net.alpha = 0.4,
  bt.size = 50,
  numCores = 7,
  CI.alpha = 0.05
)
}
\arguments{
\item{bulk.dat}{a dimension-reduced bulk data. Rows stand for samples and columns stand for PCs.}

\item{y}{class labels or survival time.
\itemize{
\item If \code{family = "binomial"} or \code{"cumulative"}, \code{y} represents class labels.
\item If \code{family = "gaussian"}, \code{y} represents a continuous variable.
\item If \code{family = "cox"}, \code{y} is a data frame with two columns. Row names of \code{y} are the same as \code{bulk.dat}'s and the two columns are named \code{("time", "status")}.
This renaming is required for the algorithm used.
}}

\item{family}{this argument takes one of the following input:
\itemize{
\item \code{family = "binomial"}: logistic regression with elastic net;
\item \code{family = "cumulative"}: reversed proportional log odds ratio model with elastic net;
\item \code{family = "gaussian"}: linear regression with elastic net; and
\item \code{family = "cox"}: cox regression with elastic net.
}}

\item{ct.res}{a list contains
\itemize{
\item \code{k}, the number of clusters;
\item \code{ct.assignment}, a data frame with one column indicating cluster assignment. Row names of \code{ct.assignment} are cell names
\item \code{centers}, cluster centroids. A data frame whose rows are for PCs and columns are for clusters. Each cluster centroid is calculated by taking the average value of all the cells in the cluster.
}}

\item{ela.net.alpha}{the parameter alpha used for the elastic net. \code{0} for ridge and \code{1} for lasso. The default is \code{ela.net.alpha = 0.4}.}

\item{bt.size}{the number of bootstrap samples. The default is \code{bt.size = 50}.}

\item{numCores}{the number of cores used for parallel computing. The default is \code{numCores = 7}.}

\item{CI.alpha}{significance level used to decide significantly positive/negative results. The default is \code{CI.alpha = 0.05}.}
}
\value{
A data frame with six columns. Row names are cells. The six columns are
\itemize{
\item (1). \code{cluster_assignment}: the cluster assignment of each cell;
\item (2). \code{Lambda.est}: the estimated Lambda;
\item (3). \code{Lambda.upper}: the upper value of the confidence interval;
\item (4). \code{Lambda.lower}: the lower value of the confidence interval;
\item (5). \code{sig}: significance identification, includes "Sig.pos", "Sig.neg", and "Not.sig";
\item (6). \code{log.pval}: log10 p-values
}
}
\description{
Obtain phenotype-associated cells using SCIPAC
}
