% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.ct.R
\name{seurat.ct}
\alias{seurat.ct}
\title{Perform the clustering method embedded in Seurat package.}
\usage{
seurat.ct(sc.dat, res = 2)
}
\arguments{
\item{sc.dat}{dimension reduced scRNA-seq data whose rows for cells and columns are for PCs.}

\item{res}{the resolution used in seurat clustering. Default is \code{res = 2.0}.}
}
\value{
a list contains
\itemize{
\item \code{k}, the number of clusters;
\item \code{ct.assignment}, a data frame with one column indicating cluster assignment. Row names of \code{ct.assignment} are cell names
\item \code{centers}, cluster centroids. Rows are for PCs and columns are for clusters. Each cluster centroid is calculated by taking the average value of all the cells in the cluster.
}
}
\description{
Perform the clustering method embedded in Seurat package.
}
